RavenDB-22462: AbstractPager includes properties and state that belongs to Tree
