+++ src/Voron/Data/BTrees/Tree.MultiTree.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Diagnostics;
 using Sparrow;
 using Sparrow.Binary;
@@ -37,9 +37,8 @@ namespace Voron.Data.BTrees
             if (!value.HasValue)
                 throw new ArgumentNullException(nameof(value));
 
-            int maxNodeSize = Llt.DataPager.NodeMaxSize;
-            if (value.Size > maxNodeSize)
-                throw new ArgumentException("Cannot add a value to child tree that is over " + maxNodeSize + " bytes in size", nameof(value));
+            if (value.Size > NodeMaxSize)
+                throw new ArgumentException("Cannot add a value to child tree that is over " + NodeMaxSize + " bytes in size", nameof(value));
             if (value.Size == 0)
                 throw new ArgumentException("Cannot add empty value to child tree");
 
@@ -52,7 +51,7 @@ namespace Voron.Data.BTrees
             var page = FindPageFor(key, out _);
             if (page == null || page.LastMatch != 0)
             {
-                MultiAddOnNewValue(key, value, maxNodeSize);
+                MultiAddOnNewValue(key, value, NodeMaxSize);
                 return;
             }
 
@@ -64,7 +63,7 @@ namespace Voron.Data.BTrees
             if (item->Flags == TreeNodeFlags.MultiValuePageRef)
             {
                 var existingTree = OpenMultiValueTree(key, item);
-                existingTree.DirectAdd(value, 0,out _).Dispose();
+                existingTree.DirectAdd(value, 0,out byte* _).Dispose();
                 return;
             }
 
@@ -105,13 +104,13 @@ namespace Voron.Data.BTrees
                 var requiredSpace = nestedPage.PageSize + // existing page
                                     nestedPage.GetRequiredSpace(value, 0); // new node
 
-                if (requiredSpace + Constants.Tree.NodeHeaderSize <= maxNodeSize)
+                if (requiredSpace + Constants.Tree.NodeHeaderSize <= NodeMaxSize)
                 {
                     // ... and it won't require to create an overflow, so we can just expand the current value, no need to create a nested tree yet
 
                     EnsureNestedPagePointer(page, item, ref nestedPage, ref nestedPagePtr);
 
-                    var newPageSize = (ushort)Math.Min(Bits.PowerOf2(requiredSpace), maxNodeSize - Constants.Tree.NodeHeaderSize);
+                    var newPageSize = (ushort)Math.Min(Bits.PowerOf2(requiredSpace), NodeMaxSize - Constants.Tree.NodeHeaderSize);
 
                     ExpandMultiTreeNestedPageSize(key, value, nestedPagePtr, newPageSize, nestedPage.PageSize);
