+++ src/Voron/Data/BTrees/TreeRebalancer.cs
@@ -75,7 +75,7 @@ namespace Voron.Data.BTrees
                 }
 
                 var minKeys = page.IsBranch ? 2 : 1;
-                if ((page.UseMoreSizeThan(_tx.DataPager.PageMinSpace)) && page.NumberOfEntries >= minKeys)
+                if ((page.UseMoreSizeThan(Tree.PageMinSpace)) && page.NumberOfEntries >= minKeys)
                     return null; // above space/keys thresholds
 
                 Debug.Assert(parentPage.NumberOfEntries >= 2); // if we have less than 2 entries in the parent, the tree is invalid
@@ -97,7 +97,7 @@ namespace Voron.Data.BTrees
                 Debug.Assert(page.IsCompressed == false);
 
                 minKeys = sibling.IsBranch ? 2 : 1; // branch must have at least 2 keys
-                if (sibling.UseMoreSizeThan(_tx.DataPager.PageMinSpace) &&
+                if (sibling.UseMoreSizeThan(Tree.PageMinSpace) &&
                     sibling.NumberOfEntries > minKeys)
                 {
                     // neighbor is over the min size and has enough key, can move just one key to  the current page
